Updated: fix interceptor signature logic on mobile
